:recycle: [Refactor] 테스트코드 디렉토리 구조 변경 반영
